# neurosentio-copilot-agent/reports/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeuroSentio Copilot Agent",
    "description": "The standalone backend brain of NeuroSentio Daily Copilot. Manages user profiles, tasks, energy logs, and generates rule-based and LLM-powered daily plans. Designed for neurodivergent-friendly pacing.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/micro-actions/{micro_action_id}/status": {
      "patch": {
        "tags": [
          "MicroActions"
        ],
        "summary": "Update micro-action status",
        "description": "Update the status of a single micro-action.\nAllowed values: open, done, snoozed, skipped, deferred.",
        "operationId": "update_micro_action_status_micro_actions__micro_action_id__status_patch",
        "parameters": [
          {
            "name": "micro_action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Micro Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MicroActionStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicroAction"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/micro-actions/{micro_action_id}/make-smaller": {
      "post": {
        "tags": [
          "MicroActions"
        ],
        "summary": "Split a micro-action into smaller actions",
        "description": "This action still feels too big.\n\nSplits it into 1\u20133 smaller actions inserted after the original in sort order.\nThe original action is NOT deleted \u2014 it stays in the list for reference.\nIf current_energy < 30, generates a lighter version (recovery mode).",
        "operationId": "make_action_smaller_micro_actions__micro_action_id__make_smaller_post",
        "parameters": [
          {
            "name": "micro_action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Micro Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MakeSmallerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakeSmallerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/copilot/next-action": {
      "get": {
        "tags": [
          "Next Action"
        ],
        "summary": "Get the best next action for the user",
        "description": "Returns exactly one best next action based on current state.\nCreates and saves a prompt row if one does not already exist.\nReuses existing active prompts where appropriate.",
        "operationId": "get_next_action_copilot_next_action_get",
        "parameters": [
          {
            "name": "at_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextActionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/copilot/next-action/{prompt_id}/done": {
      "post": {
        "tags": [
          "Next Action"
        ],
        "summary": "Mark a next action as done",
        "description": "Marks the prompt as done.\nIf source_type is micro_action, marks the linked micro-action as done too.",
        "operationId": "mark_done_copilot_next_action__prompt_id__done_post",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextActionPrompt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/copilot/next-action/{prompt_id}/snooze": {
      "post": {
        "tags": [
          "Next Action"
        ],
        "summary": "Snooze a next action",
        "description": "Snoozes the prompt for the specified number of minutes (5-1440).",
        "operationId": "snooze_action_copilot_next_action__prompt_id__snooze_post",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NextActionSnoozeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextActionPrompt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/copilot/next-action/{prompt_id}/skip": {
      "post": {
        "tags": [
          "Next Action"
        ],
        "summary": "Skip a next action",
        "description": "Marks the prompt as skipped.\nIf source_type is micro_action, optionally marks the micro-action as skipped.",
        "operationId": "skip_action_copilot_next_action__prompt_id__skip_post",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextActionPrompt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/copilot/next-action/{prompt_id}/defer": {
      "post": {
        "tags": [
          "Next Action"
        ],
        "summary": "Defer a next action to a later time",
        "description": "Defers the prompt to the specified datetime.",
        "operationId": "defer_action_copilot_next_action__prompt_id__defer_post",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NextActionDeferRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextActionPrompt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MakeSmallerRequest": {
        "properties": {
          "current_energy": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Energy"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "MakeSmallerRequest"
      },
      "MakeSmallerResponse": {
        "properties": {
          "original_micro_action": {
            "$ref": "#/components/schemas/MicroAction"
          },
          "smaller_actions": {
            "items": {
              "$ref": "#/components/schemas/MicroAction"
            },
            "type": "array",
            "title": "Smaller Actions"
          }
        },
        "type": "object",
        "required": [
          "original_micro_action",
          "smaller_actions"
        ],
        "title": "MakeSmallerResponse"
      },
...
      "MicroAction": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "parent_micro_action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Micro Action Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "energy_cost": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Cost"
          },
          "sensory_cost": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensory Cost"
          },
          "friction_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Friction Level"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "done",
              "snoozed",
              "skipped",
              "deferred"
            ],
            "title": "Status",
            "default": "open"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "MicroAction"
      },
      "MicroActionStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "done",
              "snoozed",
              "skipped",
              "deferred"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MicroActionStatusUpdate"
      },
...
      "NextActionDeferRequest": {
        "properties": {
          "defer_until": {
            "type": "string",
            "format": "date-time",
            "title": "Defer Until"
          }
        },
        "type": "object",
        "required": [
          "defer_until"
        ],
        "title": "NextActionDeferRequest"
      },
      "NextActionPrompt": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "scheduled_for": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled For"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "energy_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Cost"
          },
          "sensory_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensory Cost"
          },
          "friction_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Friction Level"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "snoozed_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snoozed Until"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "source_type",
          "action_type",
          "title",
          "message",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "NextActionPrompt"
      },
      "NextActionResult": {
        "properties": {
          "prompt": {
            "$ref": "#/components/schemas/NextActionPrompt"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "rule_based"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "reason"
        ],
        "title": "NextActionResult"
      },
      "NextActionSnoozeRequest": {
        "properties": {
          "minutes": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 5.0,
            "title": "Minutes",
            "default": 30
          }
        },
        "type": "object",
        "title": "NextActionSnoozeRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}